refactor: create SmarterMiddlewareMixin
