+++ smarter/smarter/common/classes.py
@@ -1,10 +1,12 @@
 """Common classes"""
 
+import ipaddress
 from logging import getLogger
 from typing import Any, Union
 
 import yaml
 from django.http import HttpRequest
+from django.utils.deprecation import MiddlewareMixin
 
 from smarter.common.exceptions import SmarterValueError
 from smarter.common.helpers.console_helpers import formatted_text
@@ -109,3 +111,32 @@ class SmarterHelperMixin:
                     raise SmarterValueError("String data is neither valid JSON nor YAML.") from yaml_error
         else:
             raise SmarterValueError("Unsupported data type for conversion to dict.")
+
+
+class SmarterMiddlewareMixin(MiddlewareMixin, SmarterHelperMixin):
+    """A mixin for middleware classes with helper functions."""
+
+    def get_client_ip(self, request):
+        """Get client IP address from request."""
+        # Check for real IP from various proxy headers in order of preference
+        for header in ["HTTP_X_REAL_IP", "HTTP_X_FORWARDED_FOR", "HTTP_CF_CONNECTING_IP"]:
+            ip = request.META.get(header)
+            if ip:
+                # X-Forwarded-For can contain multiple IPs, take the first (original client)
+                if header == "HTTP_X_FORWARDED_FOR":
+                    ip = ip.split(",")[0].strip()
+                # Skip internal/private IP ranges (Kubernetes pods, load balancers)
+                if not self._is_private_ip(ip.strip()):
+                    return ip.strip()
+
+        # Fallback to REMOTE_ADDR (should not be used in production behind proxies)
+        return request.META.get("REMOTE_ADDR", "127.0.0.1")
+
+    def _is_private_ip(self, ip):
+        """Check if IP is in private/internal ranges."""
+        try:
+            ip_obj = ipaddress.ip_address(ip)
+            return ip_obj.is_private or ip_obj.is_loopback or ip_obj.is_link_local
+        except ValueError:
+            # Invalid IP format
+            return True
